Perfetto plugin - Added 3 new tables: CPU frequency, Ftrace events, logcat (#24)
Added 3 new tables to the Perfetto plugin.

Logcat events - Displays all the logcat messages logged during the trace in simple linear timeline. Moved this table to a separate "Android" table group.
CPU frequency scaling - Displays the frequency chart for each CPU and how it changes. When the CPU goes idle, the CPU frequency is displayed as 0.
Ftrace events - Displays all the remaining events in the Raw table, which is a lot of the Ftrace events. These are displayed in simple timeline grouped by CPU

Also moved a couple tools that were used by multiple projects to a shared Utility project.