Perfetto plugin - Add support for an optional ProviderGuid mapping file (#25)
In Windows traces, the Provider name of each event is output along with all the event fields. Perfetto does not output the original provider name or GUID for an event. This change adds a workaround for that. It adds support for the loading of an optional XML file that has a mapping between a Provider Name and a Provider GUID. If the user then instruments their code to output the provider GUID as a debug annotation value with a specific key, the full Provider Name will be retrieved for that GUID and added as a column to the GenericEvent viewer.

The ProviderMapping.xml file needs to be placed in the plugin directory with all the binaries.

Also upped the retry count for waiting for trace_processor_shell to respond