Perfetto plugin - Got wall clock time and relative timestamps working properly (#26)
Gathered the timing information from the clock_snapshot and trace_bounds tables so that the UTC wall clock time for the start of the trace can be calculated. Needed to switch to using RelativeTimestamps for the data output events so that they can be properly aligned relative to the wall clock time. RelativeTimestamps are calculated in the source cookers.