Added build tasks to download shell. Added different names for data sources (#27)
We wanted to set up a NuGet pipeline in the Perfetto repo, but the Perfetto team does not want to maintain another packaging pipeline, even if we set it up. We could publish the package ourselves, but I don't think we want to maintain that and it seems weird for Microsoft to publish a NuGet package for Google tools.

Perfetto does publish to CIPD, GCS, and GitHub releases, so I made some VS build tasks to download a specific release from GitHub and copy the .exe we want to the the build directory.

Also added unique names for the 2 DataSources, since that messed up some scenarios in WPA (like using wpaexporter).