Added fix for missing absolute time and unknown arg field (#28)
Edge traces are missing the required UTC clocks for gathering the absolute trace start time and that caused the DataSourceInfo to not be initialized, which causes no events to be displayed. Added support to default to DateTime.Now for missing absolute trace start.

Also added a couple new value types for args.