Perfetto plugin - Fix for missing generic events (#42)
A type of Generic Events were being excluded from the previous generic event query. Generic events can come with a thread and a process, or just with a process. We were missing out on the latter. Added a query for the process_track table which allows us to query events that only have a process. Modified the giant LINQ query to use LEFT OUTER JOIN syntax because there are 2 valid paths now.

Added a table config to better view process-only Generic Events and also a table config that shows Generic Event activity.

Unit tests needed to be updated because the output changed slightly